requirements.txt
aiogram==3.3.0
boto3==1.34.11
cachetools==5.3.2
celery==5.3.6
celery-types==0.20.0
colorlog==6.8.0
dj_database_url==2.1.0
Django==5.0.1
django-axes[ipware]==6.3.0
django-cors-headers==4.3.1
django-silk==5.0.4
django-split-settings==1.2.0
django-storages==1.14.2
django_redis==5.4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.27.0
gunicorn==21.2.0
psycopg[c]==3.1.16
pyjwt==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
sentry-sdk==1.39.1
